init.qcom.usb: Set correct property in charging mode
Charging mode is invoked when executing `fastboot oem
enable-charger-screen`, for the purpose of supporting USB
charging, and the composition is set during bootup. However
not all filesystems are mounted, including the /data property
which stores persistent Android properties. The USB Framework
executes and sees that 'persist.sys.usb.config' is unset and
will default it to 'adb', which undoes the composition set by
the 'on charger' rule. Thus, simply change the 'on charger'
handling to set the 'persist.sys.usb.config' instead so that
it does not get overridden.

Change-Id: I79a61642942546806875ca67593ab7ed04affb46