init.qcom.post_boot: Remove scaling_max_freq save and restore for 8994
A57 frequency needs to be limited during boot when thermal hotplug is
disabled. We achieve this by using msm_performance module.

Remove save and restore of scaling_max_freq as it's redundant and
error-prone due to the fact that scaling_max_freq node won't be
available if a core is hotplugged out.

Change-Id: Ia1aff0b70d1e144901c6008c2725d6a6489fae91