SEEMP: Add optional package verifer to PackageManagerService
In additional to the required and sufficient package verifiers,
add support for an optional verifier in PackageManager.  The
optional verifier will act as a required verifier if a required
verifier is not found in the system.  It will act as a secondary
verifier if a required verifier is present.  Add the name of the
optional verifier as a resource string.

Change-Id: I5d53c566dd993229e43e35e9daffd35ec9f242c9