init.qcom.usb: Add changes to support DPL using IPA path
The DPL functionality may use the GSI or the IPA path
depending on the target. To select the correct path,
modify "dpl_gsi" to "dpl" in the userspace composition
script, and supply "sys.usb.rmnet.func.name" as the
function name which can be gsi or rmnet_bam depending on
the target.

Edit boot time function instance initialisation calls to
initialise a dpl function instance.

Change-Id: Iae9ea0a49fb4c22ae0ac16565b89d6f1ce635dc4