init.qcom.usb: Move setting of ConfigFS serialnumber to .sh
In some cases the 'ro.serialno' may not be populated by the
bootloader. Since init.rc syntax does not support conditional
statements, add checking of existence of this property to the
init.qcom.usb.sh script so that a dummy value can be assigned
and written to ConfigFS in case it is not defined.

Change-Id: I438316691c608056ef870e974045b547ef654409